requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0

# Data Validation & Settings
pydantic==2.9.2
pydantic-settings==2.6.0

# Configuration
python-dotenv==1.0.1

# HTTP Client
httpx==0.27.2

# LangChain & AI
langchain-core==0.3.15
langchain-google-genai==2.0.5
langgraph==0.2.52

# Google Gemini API
google-generativeai==0.8.5

# Authentication & Database
PyJWT==2.9.0
supabase==2.11.0
passlib==1.7.4
bcrypt==4.2.1
